​​refactor: setState prop 제거, 함수로 감싸서 내려주기
