+++ src/components/Product/Content/List/Item/CardItemButton/Add/index.tsx
@@ -3,15 +3,15 @@ import addCartItemIcon from "@/assets/icons/add-cart-item.svg";
 import useMutation from "@/hooks/useMutation";
 import { addCartItems } from "@/apis/cartItems/addCartItems";
 import { getCartItems } from "@/apis/cartItems/getCartItems";
-import { SetCartItems } from "@/types/cartItem";
+import { CartItemType } from "@/apis/cartItems/cartItem.type";
 import AlertToast from "@/components/AlertToast";
 
 interface AddCartItemButton {
   id: number;
-  setCartItems: SetCartItems;
+  updateCartItems: (newCartItems: CartItemType[]) => void;
 }
 
-function AddCartItemButton({ id, setCartItems }: AddCartItemButton) {
+function AddCartItemButton({ id, updateCartItems }: AddCartItemButton) {
   const { mutate, isLoading, error } = useMutation(() =>
     addCartItems({ productId: id, quantity: 1 })
   );
@@ -20,7 +20,7 @@ function AddCartItemButton({ id, setCartItems }: AddCartItemButton) {
     try {
       await mutate();
       const cartItems = await getCartItems();
-      setCartItems(cartItems);
+      updateCartItems(cartItems);
     } catch (error) {
       // mutate에서 에러가 발생한 경우 이후 로직을 실행하지 않게 try-catch문을 사용합니다.
     }
