+++ src/components/Product/Content/List/Item/CardItemButton/Remove/index.tsx
@@ -1,6 +1,6 @@
 import Button from "@/components/Button";
 import removeCartItemIcon from "@/assets/icons/remove-cart-item.svg";
-import { SetCartItems } from "@/types/cartItem";
+import { CartItemType } from "@/apis/cartItems/cartItem.type";
 import { removeCartItem } from "@/apis/cartItems/removeCartItem";
 import { getCartItems } from "@/apis/cartItems/getCartItems";
 import useMutation from "@/hooks/useMutation";
@@ -8,17 +8,17 @@ import AlertToast from "@/components/AlertToast";
 
 interface RemoveCartItemButton {
   id: number;
-  setCartItems: SetCartItems;
+  updateCartItems: (newCartItems: CartItemType[]) => void;
 }
 
-function RemoveCartItemButton({ id, setCartItems }: RemoveCartItemButton) {
+function RemoveCartItemButton({ id, updateCartItems }: RemoveCartItemButton) {
   const { mutate, isLoading, error } = useMutation(() => removeCartItem(id));
 
   const handleClick = async () => {
     try {
       await mutate();
       const cartItems = await getCartItems();
-      setCartItems(cartItems);
+      updateCartItems(cartItems);
     } catch (error) {
       // mutate에서 에러가 발생한 경우 이후 로직을 실행하지 않게 try-catch문을 사용합니다.
     }
