+++ src/pages/Product/ProductPage.tsx
@@ -2,7 +2,7 @@ import { getCartItems } from "@/apis/cartItems/getCartItems";
 import { wrapPromise } from "@/apis/wrapPromise";
 import Header from "@/components/Header";
 import ProductContent from "@/components/Product/Content";
-import { CartItemType } from "@/types/cartItem";
+import { CartItemType } from "@/apis/cartItems/cartItem.type";
 import { useState } from "react";
 
 const cartItemResource = wrapPromise(getCartItems());
@@ -11,10 +11,14 @@ function ProductPage() {
     cartItemResource.read()
   );
 
+  const updateCartItems = (newCartItems: CartItemType[]) => {
+    setCartItems(newCartItems);
+  };
+
   return (
     <>
       <Header quantity={cartItems.length} />
-      <ProductContent cartItems={cartItems} setCartItems={setCartItems} />
+      <ProductContent cartItems={cartItems} updateCartItems={updateCartItems} />
     </>
   );
 }
